requirements.txt
requests
typer
rich
img2pdf
pillow
pydantic
PyQt6
qtawesome
